// client/src/app/enums/preference-ids.enum.ts
export enum PreferenceIds {
  ShowAllObjects = 'show_all_objects',
  ShowStatic = 'show_static',
  ShowLandVehicles = 'show_land_vehicles',
  ShowAircrafts = 'show_aircrafts',
  ShowGoals = 'show_goals',
  ShowWaypoints = 'show_waypoints',
  ShowEntireWorld = 'show_entire_world',
  ShowCoordsInMeters = 'show_coords_in_meters',
  ShowDistanceBar = 'show_distance_bar',
  ShowTrueDirection = 'show_true_direction',
  ShowZoomSpot = 'show_zoom_spot',
}